+++ R/utils_print_data_frame.R
@@ -21,17 +21,17 @@ print_data_frame <- function(x, digits) {
       i,
       "conditional" = ,
       "fixed_conditional" = ,
-      "fixed" = "# fixed effects, conditional component",
+      "fixed" = " Fixed Effects (Conditional Model)",
       "zi" = ,
       "zero_inflated" = ,
       "fixed_zero_inflated" = ,
-      "fixed_zi" = "# fixed effects, zero-inflation component",
+      "fixed_zi" = " Fixed Effects (Zero-Inflated Model)",
       "random" = ,
-      "random_conditional" = "# random effects, conditional component",
+      "random_conditional" = " Random Effects (Conditional Model)",
       "random_zero_inflated" = ,
-      "random_zi" = "# random effects, zero-inflation component",
+      "random_zi" = " Random Effects (Zero-Inflated Model)",
       "smooth_sd" = ,
-      "fixed_smooth_sd" = "# smooth terms"
+      "fixed_smooth_sd" = " Smooth Terms"
     )
 
     if ("Parameter" %in% colnames(out[[i]])) {