+++ NEWS.md
@@ -1,4 +1,16 @@
-# bayestestR 0.3.3
+# bayestestR 0.4.1
+
+## General
+
+- Added `plot()`-method for `p_significance()`.
+- `p_significance()` now also works for *brmsfit*-objects.
+- `equivalence_test()` gets `effects` and `component` arguments for *stanreg* and *brmsfit*  models, to print specific model components.
+
+## Bug fixes
+
+- Fixed mistake in description of `p_significance()`.
+
+# bayestestR 0.4.0
 
 ## New functions / features